2.0 - GITHUB LIST
Welp, I finally did it, added the one and only GITHUB LIST! I can't wait to show you the release I did, working for the total 4 hours, the most I've ever worked on this project for, I'm gonna release this on to the release tab too, so stay tuned in for that but, this was a long project to do. I even added extra features such as Comments and Readmes. I hope you guys enjoy!